generate: enforce usage of the _ meta field
