+++ hosts_test.go
@@ -6,8 +6,5 @@ import (
 
 func TestListHosts(t *testing.T) {
 	req := &ListHosts{}
-	if req.name() != "listHosts" {
-		t.Errorf("API call doesn't match")
-	}
 	_ = req.response().(*ListHostsResponse)
 }
